requirements.txt
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
camelot-py==1.0.9
certifi==2026.1.4
cffi==2.0.0
chardet==6.0.0.post1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decorator==5.2.1
Deprecated==1.3.1
Django==5.2.11
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.133.1
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
narwhals==2.17.0
numpy==2.2.6
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pbr==7.0.3
pdfminer.six==20251230
pdfplumber==0.11.9
pikepdf==10.3.0
pillow==12.1.1
pluggy==1.6.0
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.27.1
pypdf==3.17.4
pypdfium2==5.5.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.47
sqlparse==0.5.5
starlette==0.52.1
streamlit==1.54.0
tabulate==0.9.0
Tempita==0.6.0
tenacity==9.1.4
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
wrapt==2.1.1
xlrd==0.7.1
xlwt==0.7.2
